requirements.txt
amqp==2.5.2
billiard==3.6.2.0
celery==4.4.0
Django==2.2.6
django-cors-headers==3.2.1
django-filter==2.2.0
django-grappelli==2.13.2
djangorestframework==3.11.0
djangorestframework-files==1.1.0
dramatiq==1.8.1
kombu==4.6.7
Markdown==3.1.1
Pillow==6.2.1
prometheus-client==0.7.1
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
vine==1.3.0
